app/app_utils/.requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1 ; python_full_version < '3.11'
attrs==25.4.0
authlib==1.6.7
certifi==2026.1.4
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==46.0.4
decorator==5.2.1
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1 ; python_full_version < '3.11'
fastapi==0.123.10
fastuuid==0.14.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
gcsfs==2026.2.0
google-adk==1.21.0
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-aiplatform==1.130.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.62.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-storage-control==1.9.0
google-cloud-texttospeech==2.34.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-genai==1.62.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.2.0 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.4.1
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.81.9
mako==1.3.10
markupsafe==3.0.3
mcp==1.26.0
mmh3==5.2.0
multidict==6.7.1
numpy==2.2.6 ; python_full_version < '3.11'
numpy==2.4.2 ; python_full_version >= '3.11'
oauthlib==3.3.1
openai==2.17.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-google-genai==0.6b0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-genai==0.2b0
packaging==26.0
pandas==2.3.3 ; python_full_version < '3.11'
pandas==3.0.0 ; python_full_version >= '3.11'
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2 ; python_full_version < '3.11'
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
ruamel-yaml==0.19.1
scikit-learn==1.5.2 ; python_full_version < '3.11'
scikit-learn==1.8.0 ; python_full_version >= '3.11'
scipy==1.15.3 ; python_full_version < '3.11'
scipy==1.17.0 ; python_full_version >= '3.11'
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.46
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.2.0
starlette==0.50.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0 ; python_full_version < '3.11'
tqdm==4.67.3
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3 ; python_full_version < '3.11' or sys_platform == 'emscripten' or sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
